HIFI-586: Make sure to await all the disconnect attempts. This was causing retries to fail immediately because the states weren't what they were expected to be, which in turn was causing oddness with signaling connections getting connected when they shouldn't be (see also HIFI-603).
Also, I changed all of the "\n"s that were embedded in template literals
to be actual newlines, because I was getting sick of seeing "\n"s
scattered all throughout the log messages. This may be a controversial
change. All I can say is that if we want to use template literals, we
have to take the good with the bad (unless someone can tell me how to
get the console output to output an actual newline instead of a \n when
it's embedded inside a template literal).